Make image slider a reuseable component
